Move test prefabs into prefabs folder
