--- Echidna2.Tests/Prefabs/SubcomponentPrefabTest.toml
@@ -1,10 +1,0 @@
-[0]
-Component = "Echidna2.TestPrefabs.SubcomponentPrefab, Echidna2.Tests"
-Subcomponent = "1"
-
-[1]
-Component = "Echidna2.TestPrefabs.SubcomponentSubcomponent, Echidna2.Tests"
-Reference = "2"
-
-[2]
-Component = "Echidna2.TestPrefabs.SubcomponentSubcomponent, Echidna2.Tests"
